Criacao de componente Box Comentario
